show loading icon during network calls
getting feeds, selecting feeds, marking entries as read, etc.
All of them should now mark a busy indicator and then hide it.
ripe for a refactor though. Would be nice to just make this CSS
applied to a busy class.

The entries loading indicator is also stuck to the top of the container
instead of being fixed to the top of the container. Fixed would be
better