proper indicator styling
make the indicators positioned ABSOLUTE, but entry positioned relative.
to fix scrolling I also eliminated a div level in entry list. We don't
neeed the entry to have an LI that does the repeat and then class
attributes on a nested div. Switched the ID and all attributes to the
LI.

Also noticed that the scroll to entry was scrolling based on commandbar
height. Commandbar isn't in the main-content anymore, so I removed it.