Added initial load for feeds and tags.
Ideally, we only pull this data once in one form and then just do
transforms on the client.
The way I'm doing it now is not great. I am trying to decide if there is
a good sql way to pull all the tags with the feeds and then group by
each tag or just pull tags with feeds and just get a unique list for
things under the feeds. both are a lot of redundant data...