Change tagging to work the right way.
Now, we pull a simple list of feeds, with tags as a csv property on each
feed.

We build a list of tags, and bucket appropriately. By doing this we save
a network call, a db call, and we are consistently using the same js
objects throughout the UI. unread_count increment decrement just works
naturally.

This is the way I should have started with.

Le sigh.